Refactor chat_logging.py and config.py
- chat_logging.py:
  - Moved chat logging functions to a separate module for better organization.
  - Added new logging functions to handle different types of chat reactions.

- config.py:
  - Centralized configuration settings for better manageability.
  - Introduced a Config class to handle environment variables and application settings.
  - Updated the application to use the new configuration class.
